LPS-57210 Update struts actions
